+++ CHANGES.rst
@@ -4,6 +4,15 @@ Changelog (Pillow)
 4.2.0 (unreleased)
 ------------------
 
+- Update Feature Detection
+  [wiredfool]
+
+- CI: Update pypy on TravisCI
+  [hugovk]
+
+- ImageMorph: Fix wrong expected size of MRLs read from disk #2561
+  [dov]
+
 - Docs: Update install docs for FreeBSD #2546
   [wiredfool]